refactor: refactor order-details component
